Make time a long, since it's time_t
* banshee/src/Core/Banshee.Services/Banshee.Collection.Indexer/CollectionIndexerService.cs:
* banshee/src/Core/Banshee.Services/Banshee.Collection.Indexer/ICollectionIndexerService.cs:
  Make time a long, since it's time_t

* banshee/src/Core/Banshee.Services/Banshee.Collection.Indexer.RemoteHelper/SimpleIndexerClient.cs:
  Support getting available export fields, and setting which fields you
  want exported to you; abstract the HasCollectionChanged by adding
  CollectionCount and CollectionLastModified abstract members that make
  the API a little less confusing

svn path=/trunk/banshee/; revision=4675

+++ src/Core/Banshee.Services/Banshee.Collection.Indexer/CollectionIndexerService.cs
@@ -136,7 +136,7 @@ namespace Banshee.Collection.Indexer
             }
         }
         
-        public bool HasCollectionLastModifiedChanged (int time)
+        public bool HasCollectionLastModifiedChanged (long time)
         {
             lock (this) {
                 long last_updated = 0;
